+++ src/bin/account.py
@@ -1,7 +1,7 @@
 import typer
 from zanshinsdk import Client
 from lib.utils import dump_json
-import config.sdk as skd_config
+import config.sdk as sdk_config
 
 ###################################################
 # Account App
@@ -15,7 +15,7 @@ def account_me():
     Returns the details of the user account that owns the API key used by this Connection instance as per
     """
     
-    client = Client(profile=skd_config.profile)
+    client = Client(profile=sdk_config.profile)
     try:
         dump_json(client.get_me())
     except Exception as e:
